Refactor to AnyServer
