+++ Src/ManiaAPI.Xml/MasterServer.cs
@@ -1,62 +1,22 @@
 ﻿using System.Collections.Immutable;
-using System.Net.Http.Headers;
 
 namespace ManiaAPI.Xml;
 
-public interface IMasterServer : IDisposable
+public interface IMasterServer : IAnyServer, IDisposable
 {
-    HttpClient Client { get; }
-
     Task<MasterServerResponse<ImmutableList<League>>> GetLeaguesResponseAsync(CancellationToken cancellationToken = default);
     Task<ImmutableList<League>> GetLeaguesAsync(CancellationToken cancellationToken = default);
     Task<MasterServerResponse<PlayerInfos>> GetPlayerInfosResponseAsync(string login, CancellationToken cancellationToken = default);
     Task<PlayerInfos> GetPlayerInfosAsync(string login, CancellationToken cancellationToken = default);
     Task<MasterServerResponse<CheckLoginResult>> CheckLoginResponseAsync(string login, CancellationToken cancellationToken = default);
     Task<CheckLoginResult> CheckLoginAsync(string login, CancellationToken cancellationToken = default);
-    Task<MasterServerResponse> TestAsync(CancellationToken cancellationToken = default);
 }
 
-public abstract class MasterServer : IMasterServer
+public abstract class MasterServer : AnyServer, IMasterServer
 {
-    public HttpClient Client { get; }
+    protected MasterServer(HttpClient client) : base(client) { }
 
-    protected abstract string GameXml { get; }
-
-    protected MasterServer(HttpClient client)
-    {
-        Client = client;
-
-        var headers = Client.DefaultRequestHeaders;
-
-        const string product = "ManiaAPI.NET";
-        const string version = "2.7.0";
-
-        var libraryExists = headers.UserAgent.Any(h => h.Product?.Name == product && h.Product?.Version == version);
-
-        if (!libraryExists)
-        {
-            headers.UserAgent.Add(new ProductInfoHeaderValue(product, version));
-            headers.UserAgent.Add(new ProductInfoHeaderValue("(Xml; Email=petrpiv1@gmail.com; Discord=bigbang1112)"));
-        }
-
-        if (client.BaseAddress is null)
-        {
-            throw new ArgumentException("BaseAddress must be set for MasterServer", nameof(client));
-        }
-    }
-
-    protected MasterServer(Uri uri) : this(new HttpClient { BaseAddress = uri }) { }
-
-    protected async Task<MasterServerResponse<T>> RequestAsync<T>(
-        string? authorXml,
-        string requestName,
-        string parametersXml,
-        XmlProcessContent<T> processContent,
-        CancellationToken cancellationToken = default) where T : notnull
-    {
-        var response = await XmlHelper.SendAsync(Client, GameXml, authorXml, requestName, parametersXml, cancellationToken);
-        return XmlHelper.ProcessResponseResult(requestName, response, processContent);
-    }
+    protected MasterServer(Uri uri) : base(uri) { }
 
     public virtual async Task<MasterServerResponse<ImmutableList<League>>> GetLeaguesResponseAsync(CancellationToken cancellationToken = default)
     {
@@ -201,18 +161,5 @@ public abstract class MasterServer : IMasterServer
     public async Task<CheckLoginResult> CheckLoginAsync(string login, CancellationToken cancellationToken = default)
     {
         return (await CheckLoginResponseAsync(login, cancellationToken)).Result;
-    }
-
-    public virtual async Task<MasterServerResponse> TestAsync(CancellationToken cancellationToken = default)
-    {
-        const string RequestName = "Test";
-        var response = await XmlHelper.SendAsync(Client, GameXml, authorXml: null, RequestName, string.Empty, cancellationToken);
-        return XmlHelper.ProcessResponseResult(response);
-    }
-
-    public virtual void Dispose()
-    {
-        Client.Dispose();
-        GC.SuppressFinalize(this);
     }
 }
