+++ Src/ManiaAPI.Xml/TMT/MasterServerTMT.cs
@@ -5,7 +5,7 @@ using TmEssentials;
 
 namespace ManiaAPI.Xml.TMT;
 
-public interface IMasterServerTMT : IMasterServer, IInitServerTMT
+public interface IMasterServerTMT : IMasterServer, IAnyServerMP
 {
     Task<MasterServerResponse<ImmutableList<SummaryZone<int>>>> GetCampaignLeaderBoardSummariesResponseAsync(IEnumerable<string> zones, CancellationToken cancellationToken = default);
     Task<MasterServerResponse<ImmutableList<SummaryZone<int>>>> GetCampaignLeaderBoardSummariesResponseAsync(string zone = "World", CancellationToken cancellationToken = default);
@@ -56,53 +56,6 @@ public class MasterServerTMT : MasterServer, IMasterServerTMT
     public async Task<WaitingParams> GetWaitingParamsAsync(string? login = null, CancellationToken cancellationToken = default)
     {
         return (await GetWaitingParamsResponseAsync(login, cancellationToken)).Result;
-    }
-
-    public virtual async Task<MasterServerResponse<string>> GetAccountFromSteamUserResponseAsync(ulong steamId, CancellationToken cancellationToken = default)
-    {
-        return await InitServerMP.GetAccountFromSteamUserResponseAsync(Client, GameXml, steamId, cancellationToken);
-    }
-
-    public async Task<string?> GetAccountFromSteamUserAsync(ulong steamId, CancellationToken cancellationToken = default)
-    {
-        try
-        {
-            return (await GetAccountFromSteamUserResponseAsync(steamId, cancellationToken)).Result;
-        }
-        catch (XmlRequestException ex)
-        {
-            if (ex.Value == 404)
-            {
-                return null;
-            }
-
-            throw;
-        }
-    }
-
-    public virtual async Task<MasterServerResponse<ImmutableList<WebIdentityPlayer>>> GetWebIdentityFromManiaplanetLoginResponseAsync(IEnumerable<string> logins, CancellationToken cancellationToken = default)
-    {
-        return await InitServerMP.GetWebIdentityFromManiaplanetLoginResponseAsync(Client, GameXml, logins, cancellationToken);
-    }
-
-    public async Task<MasterServerResponse<ImmutableList<WebIdentityPlayer>>> GetWebIdentityFromManiaplanetLoginResponseAsync(params string[] logins)
-    {
-        return await GetWebIdentityFromManiaplanetLoginResponseAsync(logins, cancellationToken: default);
-    }
-
-    public async Task<ImmutableList<WebIdentityPlayer>> GetWebIdentityFromManiaplanetLoginAsync(IEnumerable<string> logins, CancellationToken cancellationToken = default)
-    {
-        return (await GetWebIdentityFromManiaplanetLoginResponseAsync(logins, cancellationToken)).Result;
-    }
-
-    public async Task<ImmutableList<WebIdentityPlayer>> GetWebIdentityFromManiaplanetLoginAsync(params string[] logins)
-    {
-        return await GetWebIdentityFromManiaplanetLoginAsync(logins, cancellationToken: default);
-    }
-
-    public async Task<WebIdentityPlayer?> GetWebIdentityFromManiaplanetLoginAsync(string login, CancellationToken cancellationToken = default)
-    {
-        return (await GetWebIdentityFromManiaplanetLoginAsync([login], cancellationToken)).FirstOrDefault();
     }
 
     public virtual async Task<MasterServerResponse<ImmutableList<SummaryZone<int>>>> GetCampaignLeaderBoardSummariesResponseAsync(IEnumerable<string> zones, CancellationToken cancellationToken = default)
@@ -284,5 +237,40 @@ public class MasterServerTMT : MasterServer, IMasterServerTMT
         }
 
         return new Summary<T>(timestamp, units.ToImmutable());
+    }
+
+    public virtual async Task<MasterServerResponse<string>> GetAccountFromSteamUserResponseAsync(ulong steamId, CancellationToken cancellationToken = default)
+    {
+        return await AnyServerMP.GetAccountFromSteamUserResponseAsync(Client, GameXml, steamId, cancellationToken);
+    }
+
+    public virtual async Task<string?> GetAccountFromSteamUserAsync(ulong steamId, CancellationToken cancellationToken = default)
+    {
+        return await AnyServerMP.GetAccountFromSteamUserAsync(Client, GameXml, steamId, cancellationToken);
+    }
+
+    public virtual async Task<MasterServerResponse<ImmutableList<WebIdentityPlayer>>> GetWebIdentityFromManiaplanetLoginResponseAsync(IEnumerable<string> logins, CancellationToken cancellationToken = default)
+    {
+        return await AnyServerMP.GetWebIdentityFromManiaplanetLoginResponseAsync(Client, GameXml, logins, cancellationToken);
+    }
+
+    public virtual async Task<MasterServerResponse<ImmutableList<WebIdentityPlayer>>> GetWebIdentityFromManiaplanetLoginResponseAsync(params string[] logins)
+    {
+        return await AnyServerMP.GetWebIdentityFromManiaplanetLoginResponseAsync(Client, GameXml, logins);
+    }
+
+    public virtual async Task<ImmutableList<WebIdentityPlayer>> GetWebIdentityFromManiaplanetLoginAsync(IEnumerable<string> logins, CancellationToken cancellationToken = default)
+    {
+        return await AnyServerMP.GetWebIdentityFromManiaplanetLoginAsync(Client, GameXml, logins, cancellationToken);
+    }
+
+    public virtual async Task<ImmutableList<WebIdentityPlayer>> GetWebIdentityFromManiaplanetLoginAsync(params string[] logins)
+    {
+        return await AnyServerMP.GetWebIdentityFromManiaplanetLoginAsync(Client, GameXml, logins, cancellationToken: default);
+    }
+
+    public virtual async Task<WebIdentityPlayer?> GetWebIdentityFromManiaplanetLoginAsync(string login, CancellationToken cancellationToken = default)
+    {
+        return await AnyServerMP.GetWebIdentityFromManiaplanetLoginAsync(Client, GameXml, login, cancellationToken);
     }
 }
